# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "boilerplate sample API",
    "description": "sample API.",
    "termsOfService": "https://github.com/jhwmon/express-boilerplate-mysql",
    "contact": {
      "email": "jmonaco88@gmail.com"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
...
  "paths": {
...
    "/adpartners/{id}": {
      "get": {
        "tags": ["AdPartner"],
        "summary": "Retrieve user by id",
        "operationId": "findUserById",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "access token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 2000
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "adpartner": {
                          "type": "object",
                          "properties": {
                            "no": {
                              "type": "integer",
                              "example": 100
                            },
                            "userid": {
                              "type": "string",
                              "example": "jmonaco88@bbangyanetworks.com"
                            },
                            "nickname": {
                              "type": "string",
                              "example": "빵야빵야"
                            },
                            "name": {
                              "type": "string",
                              "example": "본명"
                            },
                            "level": {
                              "type": "integer",
                              "example": 1
                            },
                            "phone": {
                              "type": "string",
                              "example": "01011112222"
                            },
                            "email": {
                              "type": "string",
                              "example": "jmonaco88@bbangyanetworks.com"
                            },
                            "reward": {
                              "type": "number",
                              "example": "10.5"
                            },
                            "bankName": {
                              "type": "string",
                              "example": "기업은행"
                            },
                            "bankOwner": {
                              "type": "string",
                              "example": "통장 소유명"
                            },
                            "bankAccount": {
                              "type": "string",
                              "example": "1111-111-2222"
                            },
                            "bankImageUrl": {
                              "type": "string",
                              "example": "http://image1.jpg"
                            },
                            "createTime": {
                              "type": "string",
                              "example": "2021-10-22 12:00:00"
                            },
                            "updateTime": {
                              "type": "string",
                              "example": "2021-10-22 13:01:02"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing credentials",
            "content": {}
          },
          "401": {
            "description": "Invalid token, please log in or sign up",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 4010
                    },
                    "msg": {
                      "type": "string",
                      "example": "인증 오류"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        },
                        "trace": {
                          "type": "object",
                          "properties": {
                            "statusCode": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-codegen-request-body-name": "body"
      },
      "patch": {
        "tags": ["AdPartner"],
        "summary": "Update User By Id",
        "operationId": "updateUserById",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "access token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "User fields to update",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "reward": {
                    "type": "number"
                  },
                  "bankName": {
                    "type": "string"
                  },
                  "bankOwner": {
                    "type": "string"
                  },
                  "bankAccount": {
                    "type": "string"
                  },
                  "bankImageFile": {
                    "type": "string",
                    "format": "binary"
                  },
                  "deleteBankImage": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 2000
                    },
                    "msg": {
                      "type": "string",
                      "example": "수정 성공"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "adpartner": {
                          "type": "object",
                          "properties": {
                            "no": {
                              "type": "integer",
                              "example": 100
                            },
                            "userid": {
                              "type": "string",
                              "example": "jmonaco88@bbangyanetworks.com"
                            },
                            "nickname": {
                              "type": "string",
                              "example": "빵야빵야"
                            },
                            "name": {
                              "type": "string",
                              "example": "본명"
                            },
                            "level": {
                              "type": "integer",
                              "example": 1
                            },
                            "phone": {
                              "type": "string",
                              "example": "01011112222"
                            },
                            "email": {
                              "type": "string",
                              "example": "jmonaco88@bbangyanetworks.com"
                            },
                            "reward": {
                              "type": "number",
                              "example": "10.5"
                            },
                            "bankName": {
                              "type": "string",
                              "example": "기업은행"
                            },
                            "bankOwner": {
                              "type": "string",
                              "example": "통장 소유명"
                            },
                            "bankAccount": {
                              "type": "string",
                              "example": "1111-111-2222"
                            },
                            "bankImageUrl": {
                              "type": "string",
                              "example": "http://image1.jpg"
                            },
                            "createTime": {
                              "type": "string",
                              "example": "2021-10-22 12:00:00"
                            },
                            "updateTime": {
                              "type": "string",
                              "example": "2021-10-22 13:01:02"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing credentials",
            "content": {}
          },
          "401": {
            "description": "Invalid token, please log in or sign up",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        },
                        "trace": {
                          "type": "object",
                          "properties": {
                            "statusCode": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-codegen-request-body-name": "body"
      },
      "delete": {
        "tags": ["AdPartner"],
        "summary": "Delete User By Id",
        "operationId": "deleteUserById",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "access token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing credentials",
            "content": {}
          },
          "401": {
            "description": "Invalid token, please log in or sign up",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        },
                        "trace": {
                          "type": "object",
                          "properties": {
                            "statusCode": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-codegen-request-body-name": "body"
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}